refactor: bootstrap nest application in separate module
details: clearly split server (listening to a port) and serverless implementations
