+++ src/bootstrap.ts
@@ -1,0 +1,20 @@
+import { NestFactory } from '@nestjs/core';
+import helmet from 'helmet';
+
+import { AppModule } from './app.module';
+import { ValidationPipe } from '@nestjs/common';
+
+async function bootstrap() {
+  const app = await NestFactory.create(AppModule);
+
+  app.enableCors({
+    origin: (req, callback) => callback(null, true),
+  });
+  app.use(helmet());
+  app.useGlobalPipes(new ValidationPipe());
+  await app.init();
+
+  return app;
+}
+
+export { bootstrap };
