+++ src/handlers/cart.ts
@@ -1,6 +1,6 @@
 import serverlessExpress from '@codegenie/serverless-express';
 import { Callback, Context, Handler } from 'aws-lambda';
-import { bootstrap } from '../main.js';
+import { bootstrap } from '../bootstrap.js';
 
 let server: Handler;
 
@@ -14,6 +14,8 @@ export const handler: Handler = async (
   context: Context,
   callback: Callback,
 ) => {
+  console.log('received event:', event)
+  console.log('reeived context:', context)
   server = server ?? (await bootstrapServer());
   return server(event, context, callback);
 };
